test gitIgnore file

+++ thinker/src/main/java/com/likelion/thinker/service/BookMarkService.java
@@ -6,6 +6,7 @@ import com.likelion.thinker.entity.Member;
 import com.likelion.thinker.repository.BookMarkRepository;
 import com.likelion.thinker.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,21 +16,41 @@ public class BookMarkService {
     private final MemberRepository memberRepository;
     private final BookMarkRepository bookMarkRepository;
 
+    @Value("${cloud.accessKey}")
+    private String accessKey;
+
+    @Value("${cloud.secretKey}")
+    private String secretKey;
+
+    @Value("{cloud.bucketName}")
+    private String bucketName;
+
     @Transactional
     public String addBookMark(BookMarkDto bookMarkDto) {
         Member member = memberRepository.findByMemberId(bookMarkDto.getUserId());
-        BookMark bookMark = bookMarkRepository.save(BookMark.toBookMark(bookMarkDto, member));
+        BookMark tempBookMark = bookMarkRepository.findByBookMarkIdAndMemberId(member.getMemberId(), bookMarkDto.getUrl());
 
-        String responseId = bookMark.getUrl();
+        if(tempBookMark == null) {
+            BookMark bookMark = bookMarkRepository.save(BookMark.toBookMark(bookMarkDto, member));
+            String responseId = bookMark.getUrl();
 
-        return responseId;
+            return responseId;
+        } else {
+            return null;
+        }
     }
 
     @Transactional
     public String deleteBookMak(BookMarkDto bookMarkDto) {
         Member member = memberRepository.findByMemberId(bookMarkDto.getUserId());
-        bookMarkRepository.delete(BookMark.toBookMark(bookMarkDto, member));
+        BookMark bookMark = bookMarkRepository.findByBookMarkIdAndMemberId(member.getMemberId(), bookMarkDto.getUrl());
 
-        return bookMarkDto.getUrl();
+        if(bookMark == null) {
+            return null;
+        } else {
+            bookMarkRepository.delete(BookMark.toBookMark(bookMarkDto, member));
+
+            return bookMarkDto.getUrl();
+        }
     }
 }
